refactor tests, move things around
