+++ api_tests/providers/preprints/views/test_preprint_provider_list.py
@@ -10,13 +10,10 @@ from osf_tests.factories import (
 @pytest.mark.django_db
 class TestPreprintProviderList:
 
-    @pytest.fixture()
-    def url(self):
-        return '/{}preprint_providers/?version=2.2&'.format(API_BASE)
-
-    @pytest.fixture()
-    def url_generalized(self):
-        return '/{}providers/preprints/?version=2.2&'.format(API_BASE)
+    @pytest.fixture(params=['/{}preprint_providers/?version=2.2&', '/{}providers/preprints/?version=2.2&'])
+    def url(self, request):
+        url = (request.param)
+        return url.format(API_BASE)
 
     @pytest.fixture()
     def user(self):
@@ -64,34 +61,5 @@ class TestPreprintProviderList:
             provider_one, provider_two):
         res = app.get('{}filter[{}]={}'.format(
             url, filter_type, filter_value))
-        assert res.status_code == 200
-        assert len(res.json['data']) == 1
-
-    def test_preprint_provider_list_for_generalized_endpoint(
-            self, app, url_generalized, user, provider_one, provider_two):
-        # Test length and not auth
-        res = app.get(url_generalized)
-        assert res.status_code == 200
-        assert len(res.json['data']) == 2
-
-        # Test length and auth
-        res = app.get(url_generalized, auth=user.auth)
-        assert res.status_code == 200
-        assert len(res.json['data']) == 2
-
-    @pytest.mark.parametrize('filter_type,filter_value', [
-        ('allow_submissions', True),
-        ('description', 'spots%20not%20dots'),
-        ('domain', 'https://www.spotarxiv.com'),
-        ('domain_redirect_enabled', True),
-        ('id', 'spot'),
-        ('name', 'Spotarxiv'),
-        ('share_publish_type', 'Thesis'),
-    ])
-    def test_preprint_provider_list_filtering_for_generalized_endpoint(
-            self, filter_type, filter_value, app, url_generalized,
-            provider_one, provider_two):
-        res = app.get('{}filter[{}]={}'.format(
-            url_generalized, filter_type, filter_value))
         assert res.status_code == 200
         assert len(res.json['data']) == 1
