+++ api_tests/providers/preprints/views/test_preprint_provider_moderator_list.py
@@ -18,13 +18,10 @@ class TestPreprintProviderModeratorList:
         GroupHelper(pp).update_provider_auth_groups()
         return pp
 
-    @pytest.fixture()
-    def url(self, provider):
-        return '/{}preprint_providers/{}/moderators/'.format(API_BASE, provider._id)
-
-    @pytest.fixture()
-    def url_generalized(self, provider):
-        return '/{}providers/preprints/{}/moderators/'.format(API_BASE, provider._id)
+    @pytest.fixture(params=['/{}preprint_providers/{}/moderators/', '/{}providers/preprints/{}/moderators/'])
+    def url(self, provider, request):
+        url = request.param
+        return url.format(API_BASE, provider._id)
 
     @pytest.fixture()
     def admin(self, provider):
@@ -125,78 +122,6 @@ class TestPreprintProviderModeratorList:
     def test_list_post_admin_success_email(self, mock_mail, app, url, nonmoderator, moderator, admin, provider):
         payload = self.create_payload(email='somenewuser@gmail.com', full_name='Some User', permission_group='moderator')
         res = app.post_json_api(url, payload, auth=admin.auth)
-        assert res.status_code == 201
-        assert len(res.json['data']['id']) == 5
-        assert res.json['data']['attributes']['permission_group'] == 'moderator'
-        assert 'email' not in res.json['data']['attributes']
-        assert mock_mail.call_count == 1
-
-    def test_list_get_not_authorized_for_generalized_endpoint(self, app, url_generalized, nonmoderator, admin, provider):
-        res = app.get(url_generalized, expect_errors=True)
-        assert res.status_code == 401
-
-        res = app.get(url_generalized, auth=nonmoderator.auth, expect_errors=True)
-        assert res.status_code == 403
-
-    def test_list_get_moderator_for_generalized_endpoint(self, app, url_generalized, nonmoderator, moderator, admin, provider):
-        # admin/nonmoderator unused here, just len verification
-        res = app.get(url_generalized, auth=moderator.auth)
-        assert res.status_code == 200
-        assert len(res.json['data']) == 2
-        for datum in res.json['data']:
-            if datum['id'] == moderator._id:
-                assert datum['attributes']['permission_group'] == 'moderator'
-
-    def test_list_get_admin_with_filter_for_generalized_endpoint(self, app, url_generalized, nonmoderator, moderator, admin, provider):
-        # mod/nonmoderator unused here, just len verification
-        res = app.get('{}?filter[permission_group]=admin&filter[id]={}'.format(url_generalized, admin._id), auth=admin.auth)
-        assert res.status_code == 200
-        assert len(res.json['data']) == 1
-        assert res.json['data'][0]['id'] == admin._id
-        assert res.json['data'][0]['attributes']['permission_group'] == 'admin'
-
-    @mock.patch('framework.auth.views.mails.send_mail')
-    def test_list_post_unauthorized_for_generalized_endpoint(self, mock_mail, app, url_generalized, nonmoderator, moderator, provider):
-        payload = self.create_payload(user_id=nonmoderator._id, permission_group='moderator')
-        res = app.post(url_generalized, payload, expect_errors=True)
-        assert res.status_code == 401
-
-        res = app.post(url_generalized, payload, auth=nonmoderator.auth, expect_errors=True)
-        assert res.status_code == 403
-
-        res = app.post(url_generalized, payload, auth=moderator.auth, expect_errors=True)
-        assert res.status_code == 403
-
-        assert mock_mail.call_count == 0
-
-    @mock.patch('framework.auth.views.mails.send_mail')
-    def test_list_post_admin_success_existing_user_for_generalized_endpoint(self, mock_mail, app, url_generalized, nonmoderator, moderator, admin, provider):
-        payload = self.create_payload(user_id=nonmoderator._id, permission_group='moderator')
-
-        res = app.post_json_api(url_generalized, payload, auth=admin.auth)
-        assert res.status_code == 201
-        assert res.json['data']['id'] == nonmoderator._id
-        assert res.json['data']['attributes']['permission_group'] == 'moderator'
-        assert mock_mail.call_count == 1
-
-    @mock.patch('framework.auth.views.mails.send_mail')
-    def test_list_post_admin_failure_existing_moderator_for_generalized_endpoint(self, mock_mail, app, url_generalized, moderator, admin, provider):
-        payload = self.create_payload(user_id=moderator._id, permission_group='moderator')
-        res = app.post_json_api(url_generalized, payload, auth=admin.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert mock_mail.call_count == 0
-
-    @mock.patch('framework.auth.views.mails.send_mail')
-    def test_list_post_admin_failure_invalid_group_for_generalized_endpoint(self, mock_mail, app, url_generalized, nonmoderator, moderator, admin, provider):
-        payload = self.create_payload(user_id=nonmoderator._id, permission_group='citizen')
-        res = app.post_json_api(url_generalized, payload, auth=admin.auth, expect_errors=True)
-        assert res.status_code == 400
-        assert mock_mail.call_count == 0
-
-    @mock.patch('framework.auth.views.mails.send_mail')
-    def test_list_post_admin_success_email_for_generalized_endpoint(self, mock_mail, app, url_generalized, nonmoderator, moderator, admin, provider):
-        payload = self.create_payload(email='somenewuser@gmail.com', full_name='Some User', permission_group='moderator')
-        res = app.post_json_api(url_generalized, payload, auth=admin.auth)
         assert res.status_code == 201
         assert len(res.json['data']['id']) == 5
         assert res.json['data']['attributes']['permission_group'] == 'moderator'
